pyproject.toml
[project]
name = "fake-py-mcp"
description = "MCP server for fake.py."
readme = "README.rst"
version = "0.1"
dependencies = []
authors = [
    {name = "Artur Barseghyan", email = "artur.barseghyan@gmail.com"},
]
maintainers = [
    {name = "Artur Barseghyan", email = "artur.barseghyan@gmail.com"},
]
license = "MIT"

classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python",
    "Topic :: Software Development :: Testing",
    "Topic :: Software Development",
]
keywords = ["fake data", "test data"]

[project.urls]
Homepage = "https://github.com/barseghyanartur/fake-py-mcp/"
Issues = "https://github.com/barseghyanartur/fake-py-mcp/issues"
Documentation = "https://fake-py-mcp.readthedocs.io/"
Repository = "https://github.com/barseghyanartur/fake-py-mcp/"
Changelog = "https://fake-py-mcp.readthedocs.io/en/latest/changelog.html"

[project.optional-dependencies]
all = ["fake.py[dev,test,docs]"]
dev = [
    "build",
    "detect-secrets",
    "doc8",
    "ipython",
    "mypy",
    "pkginfo",
    "pydoclint",
    "ruff",
    "twine",
    "uv",
    "wheel",
]
test = [
    "pytest",
    "pytest-cov",
    "pytest-codeblock",
]
docs = [
    "sphinx<6.0",
    "sphinx-autobuild",
    "sphinx-rtd-theme>=1.3.0",
    "sphinx-no-pragma",
]

[project.scripts]
fake-py-mcp = "fakepy_mcp:main"

[tool.setuptools]
py-modules = ["fakepy_mcp"]

[tool.setuptools.packages.find]
where = ["."]
include = ["fakepy_mcp"]
exclude = [".github", "docs*", "examples*"]

# Build system information below.
[build-system]
requires = ["setuptools>=41.0", "setuptools-scm", "wheel"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 80

# Enable Pyflakes `E` and `F` codes by default.
lint.select = [
#    "ANN",  # Annotations: missing (return) types
    "B",  # Bugbear: common bug patterns or design problems
    "C4",  # Complexity: McCabe complexity checker
    "E",  # Pycodesyle: Style guide enforcement
    "F",  # Pyflakes: Source files errors
    "G",  # Logging format: Basic logging format errors
    "I",  # Isort: import sorting
    "ISC",  # Naming: Variable naming convention
    "INP",  # Implicit namespace packages: Check if __init__.py is present
    "N",  # Naming: Variable naming convention
    "PERF",  # Perflint: Performance linting
    "Q",
    "SIM",  # Simplify: Helps simplify your code
    "TD",  # TO DO: Format TO DO comments
]
lint.ignore = [
    "G004",  # Allow use of f-string in logging
    "ISC003",  # Allow explicitly concatenated strings
#    "N805",  # Allow first argument of a method be named other than `self`
    "TD002",  # Allow to do without author
    "TD003",  # Allow to do without URL
]
# Enable autofix for formatting and import sorting
fix = true
src = [
]

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    "docs",
]

# Allow unused variables when underscore-prefixed.
lint.dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

# Assume Python 3.9.
target-version = "py39"

[tool.ruff.lint.per-file-ignores]
"conftest.py" = [
    "PERF203"  # Allow `try`-`except` within a loop incurs performance overhead
]

[tool.ruff.lint.isort]
known-first-party = [
]
known-third-party = []

[tool.doc8]
ignore-path = [
    "docs/requirements.txt",
    "fake-py-mcp.py.egg-info/SOURCES.txt",
    "fake_py_mcp.py.egg-info/SOURCES.txt",
    "docs/llms.rst",
]

[tool.pytest.ini_options]
addopts = [
    "-ra",
    "-vvv",
    "-q",
   "--cov=fakepy_mcp",
   "--ignore=.tox",
   "--ignore=requirements",
   "--ignore=release",
   "--ignore=examples",
   "--ignore=tmp",
   "--cov-report=html",
   "--cov-report=term",
   "--cov-report=annotate",
   "--cov-append",
   "--capture=no",
]
testpaths = [
    "fakepy_mcp.py",
    "*.rst",
    "**/*.rst",
]
pythonpath = [
    "examples/django",
]
norecursedirs = [
    ".git",
    "examples",
]

[tool.coverage.run]
relative_files = true
omit = [
    ".tox/*",
]

[tool.coverage.report]
show_missing = true
exclude_lines = [
    "pragma: no cover",
    "@overload",
]

[tool.mypy]
check_untyped_defs = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true

[tool.pydoclint]
style = 'sphinx'
exclude = '\.git|\.tox|tests/data'
require-return-section-when-returning-nothing = false
allow-init-docstring = true
arg-type-hints-in-docstring = false
